WIP: Update reprinting type used in reprintingSort
This now takes the zipper rather than just the an out of context node
allowing reprinting functions to take into account surrounding context.

Added a somewhat contrived example test where it only adds comments if
the variable is not via a literal, which would render the comment
pointless. This could be achieved by just looking at the decl in this
case, but in more complex languages printing of say brackets might be
dependent on context.